feat(products): per-product VAT rate (0/6/12) as Float + searchable MXIK in POS
- vatRate Int → Float in both Prisma schemas (+ PG migration, SQLite REAL column,
  regenerated clients); DTOs IsInt → IsNumber.
- Product forms (web + renderer): VAT select with 0.00 / 6.00 / 12.00 + "use store
  default"; parseFloat. New products default to store default (not hardcoded 0%).
- Product lists (web + renderer): show VAT % column.
- Renderer ProductForm: searchable MXIK picker (mirrors web) via new mxik:catalogSearch
  IPC that proxies tasnif elasticsearch from the main process (no CORS).
- Fiscal heal: probe 6% too when correcting a rejected rate.

+++ src/main/database/sqlite-client.ts
@@ -609,9 +609,11 @@ async function runMigrations(prisma: PrismaClientType): Promise<void> {
     )
   `;
 
-  // Migration 24: per-product VAT rate (% — e.g. 0 or 12). Null → use global regos_vcr_vat default.
+  // Migration 24: per-product VAT rate (% — e.g. 0, 6 or 12). Null → use global regos_vcr_vat default.
+  // REAL so fractional rates round-trip; SQLite type affinity stores whole rates (0/6/12) losslessly
+  // and existing terminals (column already INTEGER affinity) handle these whole values fine.
   if (!(await columnExists(prisma, 'products', 'vat_rate'))) {
-    await prisma.$executeRaw`ALTER TABLE products ADD COLUMN vat_rate INTEGER`;
+    await prisma.$executeRaw`ALTER TABLE products ADD COLUMN vat_rate REAL`;
   }
 }
